feat(WIP): instantiate templateset to generate app release
