+++ bcs-app/backend/apps/templatesets/models/release.py
@@ -15,10 +15,15 @@ from dataclasses import dataclass
 from typing import List
 
 from django.db import models
+from django.utils import timezone
 from jsonfield import JSONField
 
 from backend.apps.configuration import models as config_models
 from backend.utils.models import BaseModel, BaseTSModel
+
+
+def gen_time_revision() -> str:
+    return timezone.localtime().strftime('%Y%m%d%H%M%S')
 
 
 @dataclass
@@ -29,51 +34,67 @@ class ResourceManifest:
     body: str
 
 
+@dataclass
+class AppReleaseData:
+    name: str
+    cluster_id: str
+    namespace: str
+    show_version: config_models.ShowVersion
+    manifest_list: List[ResourceManifest]
+
+
 class AppRelease(BaseModel):
     name = models.CharField(max_length=128)
     cluster_id = models.CharField(max_length=32)
     namespace = models.CharField(max_length=64)
-    # 关联模板集
+    status = models.CharField()
+    message = models.TextField()
     template_id = models.IntegerField("关联model Template")
-    rel_revision = models.CharField("release的修订版号", max_length=32)
-    rel_revision_history = JSONField(default=[])
+    revision = models.CharField("release的修订版号", max_length=32)
+    revision_history = JSONField(default=[])
 
     class Meta:
         db_table = 'templatesets_app_release'
         unique_together = ('name', 'cluster_id', 'namespace')
 
     def save(self, *args, **kwargs):
-        self.rel_revision_history.append(self.rel_revision)
+        self.revision_history.append(self.revision)
         return super().save(*args, **kwargs)
 
     @property
     def template(self):
         return config_models.Template.objects.get(id=self.template_id)
 
-    def deploy(self, version: str, revision: str, manifest_list: List[ResourceManifest]):
-        for manifest in manifest_list:
-            try:
-                ResourceInstance.objects.create(
-                    app_release=self,
-                    rel_revision=self.rel_revision,
-                    kind=manifest.kind,
-                    version=version,
-                    revision=revision,
-                    manifest=manifest.body,
-                )
-            except Exception:
-                pass
+    @classmethod
+    def create(cls, release_data: AppReleaseData):
+        show_version = release_data.show_version
+        app_release = cls(
+            name=release_data.name,
+            cluster_id=release_data.cluster_id,
+            namespace=release_data.namespace,
+            template_id=show_version.template_id,
+            rel_revision=gen_time_revision(),
+        )
+
+    # def deploy(self, version: str, revision: str, manifest_list: List[ResourceManifest]):
+    #     for manifest in manifest_list:
+    #         try:
+    #             ResourceInstance.objects.create(
+    #                 app_release=self,
+    #                 rel_revision=self.rel_revision,
+    #                 kind=manifest.kind,
+    #                 version=version,
+    #                 revision=revision,
+    #                 manifest=manifest.body,
+    #             )
+    #         except Exception:
+    #             pass
 
     def update(self):
         pass
 
     def delete(self, *args, **kwargs):
         pass
-
-
-# class ReleaseRevision(BaseModel):
-#     app_release = models.ForeignKey(AppRelease, on_delete=models.CASCADE, null=False)
-#     version_suffix = models.CharField("区分小版本的后缀", max_length=32)
 
 
 class ResourceInstance(BaseTSModel):
@@ -86,14 +107,16 @@ class ResourceInstance(BaseTSModel):
     version = models.CharField("模板集版本名", max_length=255)
     revision = models.CharField("模板集版本名的修订版号", max_length=32)
     is_edited = models.BooleanField("是否在线编辑过", default=False)
+    status = models.CharField()
 
     class Meta:
         db_table = 'templatesets_resource_instance'
 
     @classmethod
-    def deploy(cls):
+    def create(cls):
         # TODO 利用dynamic client部署到集群中
         pass
 
-    def copy(self, rel_revision):
+    @classmethod
+    def copy(cls, rel_revision):
         return ResourceInstance.objects.create()
